+++ src/com/kaylerrenslow/armaDialogCreator/gui/canvas/api/ControlListChangeType.java
@@ -13,6 +13,6 @@ package com.kaylerrenslow.armaDialogCreator.gui.canvas.api;
 /**
  Created by Kayler on 08/17/2016.
  */
-public enum ChangeType {
+public enum ControlListChangeType {
 	ADD, SET, REMOVE, MOVE
 }