+++ src/com/kaylerrenslow/armaDialogCreator/gui/img/Images.java
@@ -16,7 +16,8 @@ import javafx.scene.image.Image;
  Created by Kayler on 10/07/2016.
  */
 public class Images {
-	public static final Image IMAGE_ADC_ICON = new Image(ImagePaths.ICON_APP);
-
-	public static final Image IMAGE_ICON_FOLDER = new Image(ImagePaths.ICON_FOLDER);
+	public static final Image ICON_ADC = new Image(ImagePaths.ICON_APP);
+	public static final Image ICON_FOLDER = new Image(ImagePaths.ICON_FOLDER);
+	public static final Image ICON_UNDO = new Image(ImagePaths.ICON_UNDO);
+	public static final Image ICON_REDO = new Image(ImagePaths.ICON_REDO);
 }